feat: load, write supports iceberg (#3737)
* feat: load/write supports iceberg

* docs
