+++ docs/en/integration/offline_data_sources/index.rst
@@ -7,3 +7,4 @@ Offline Data Source
 
     hive
     s3
+    iceberg
