+++ docs/zh/integration/offline_data_sources/index.rst
@@ -7,3 +7,4 @@
 
     hive
     s3
+    iceberg
